requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.1
openmeteo-requests==1.3.0
psycopg2-binary==2.9.10
flask-cors==5.0.0
cryptography==37.0.2
PyJWT==2.8.0